requirements.txt
streamlit
pandas
plotly
python-dotenv
mysql-connector-python
langchain
langchain-experimental
langchain-groq
